# [WIP] Implement lazy scaling and shifting for VectorDict