Draft PR related to [this](https://github.com/creme-ml/creme/pull/338#issuecomment-689853117) and [that](https://github.com/creme-ml/creme/pull/316#issuecomment-692907210) comment, implements lazy scaling and shifting for `VectorDict`.
First 2 commits are preparation, just look at the 3rd one for the implementation.

Benchmarks below use `.keys()` to force the evaluation of the lazy part so it's comparable to the non-lazy implementation (copy at each operation).

```py
from creme import utils
x = {i: i for i in range(100)}
y = utils.VectorDict(x)

# this PR
%timeit y.keys()
106 ns ± 0.0774 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
%timeit y + 1
139 ns ± 0.337 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
%timeit (y + 1).keys()
6 µs ± 32.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
%timeit (y ** 2 / 2 + 1).keys()
31.1 µs ± 107 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
%timeit (2 * y + 3 * y).keys()
14.2 µs ± 42.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

# previous PR #351, should be similar to master
%timeit y.keys()
108 ns ± 0.132 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
%timeit y + 1
5.08 µs ± 20.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
%timeit (y + 1).keys()
5.13 µs ± 12.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
%timeit (y ** 2 / 2 + 1).keys()
35.6 µs ± 106 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
%timeit (2 * y + 3 * y).keys()
20.5 µs ± 86.5 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

We can see it's a bit faster for complex operations but it's not huge and it makes other operations a bit slower.
Also it hurts maintainability.

Maybe some custom methods as mentioned [here](https://github.com/creme-ml/creme/pull/316#issuecomment-658676077) would be a better option although a wrapper would be needed to get the same expressions with `numpy` arrays.