@MaxHalford 2020-09-18
Thanks Bruno. 

As you hinted at, I don't think that there's more cons than pros. It is an interesting topic nonetheless, and certainly one where I should spend more time looking at. In a way, numpy sort of has the same problem, right? That's why [numexpr](https://github.com/pydata/numexpr) was created. Maybe that if we grow bigger, there might some justification to spend more time here and dedicate effort towards a tidy module that is dedicated to `VectorDict`. I don't know.

But anyway, things are very good right now, no need to throw the kitchen sink at the problem, at least that's how things seem to me.