@brcharron 2020-09-18
Indeed `numpy` has the same problem although since it's pretty fast already it's a bit less of an issue. I will close this one then and maybe this can be revisited in the future with an option that can deal with both arrays and dictionaries, potentially on the lines of `numexpr`.